+++ source/plwin.h
@@ -1,2 +1,2 @@
-void matface(int *par,int *nin,double *in_r,double *in_i,
+void matface(int *par,int *nin,short *in_r,short *in_i,
 	int *nout, double *out_r, double *out_i, double *upar);
